refactor: update imports to use shadcn directory in stock adjustments pages
